feat: DashBoard task 바로 반영되도록 수정
