+++ project-management/src/types/task.ts
@@ -1,0 +1,12 @@
+export interface Task {
+  id: number;
+  title: string;
+  description: string;
+  startDate: string;
+  endDate: string;
+  teamMemberId: number;
+}
+
+export interface Tasks{
+  tasks:Task[];
+}
